+++ test/test_helper.rb
@@ -46,6 +46,7 @@ class CreateUsers < TestMigration
       t.string :city
       t.decimal :credit_score, precision: 10, scale: 5
       t.timestamp :deleted_at
+      t.string :country, limit: 20
       t.references :order
     end